chatlm-streamlit/app/chatbot_app.py
from datetime import datetime
from io import StringIO
import streamlit as st
import openai
import dotenv
import os, logging
import markdown as md
from markdown.extensions.codehilite import CodeHiliteExtension
from utils import make_chat_container, on_send_button_clicked, at_add_system_msg_button_clicked

dotenv.load_dotenv(override=True)



system_msg = """
The following is a conversation with a helpful AI assistant. The assistant is helpful, creative, clever, and very friendly. The current date is {today}.
""".strip()

# get chat_history from st cache
@st.cache_resource
def get_chat_history():
    today = datetime.now().strftime("%Y-%m-%d")
    chat_history = []
    chat_history.append({'role': 'system', 'content': system_msg.format(today=today)})
    return chat_history

# APP ==================================================

st.set_page_config(page_title="ChatLM Streamlit App", page_icon=":robot_face:")
st.title("ChatLM Streamlit App")
# st.subheader("Chat with our Chatbot")

logging.basicConfig(level=logging.DEBUG)
logger = logging.getLogger()
formatter = logging.Formatter('(%(name)s) %(asctime)s - %(levelname)s: %(message)s')

@st.cache_resource
def get_log_stream():
    log_stream = StringIO()
    logger.setLevel(logging.INFO)
    handler = logging.StreamHandler(log_stream)
    handler.setFormatter(formatter)
    logger.addHandler(handler)
    # change format of log messages
    return log_stream

log_stream = get_log_stream()

chat_area = st.empty()
user_input = st.text_area("Type your message to the Chatbot:", placeholder="Type your message here")

# Injecting CSS style for scrollbar appearance
st.markdown("""
<style>
::-webkit-scrollbar {
  width: 6px;
  height: 6px;
}

::-webkit-scrollbar-thumb {
  background-color: #6c757d;
  border-radius: 4px;
}
</style>
""", unsafe_allow_html=True)


#script to scroll to the bottom of the chat area
st.markdown("""
<script>
function scrollToBottom() {
    var chat_area = document.getElementsByClassName('scrollable-container')[0];
    chat_area.scrollTop = chat_area.scrollHeight;
}
</script>
""", unsafe_allow_html=True)


chat_history = get_chat_history()

# sidebar to append system messages
with st.sidebar:
    default_model = os.getenv("CHAT_MODEL")
    # dropdown to select the chatbot model
    model = st.selectbox(
        "Select a Chatbot Model",
        ["gpt-4", "gpt-3.5-turbo"],
        index=0 if default_model == "gpt4" else 1,
    )
    # checkbox to toggle the system messages
    show_system_msg = st.checkbox("Show System Messages", value=False)
    # text area to allow the user write system messages
    system_msg_area = st.text_area("System Messages", system_msg, placeholder="You are ...", height=100)
    # button to update system messages
    
    st.button("Add", on_click=at_add_system_msg_button_clicked, args=(system_msg_area, chat_history))

# print("Chat history:", chat_history)
openai.api_key = os.getenv("OPENAI_API_KEY")

send_button = st.button("Send", on_click=on_send_button_clicked, args=(chat_area, user_input, chat_history, model))


if not show_system_msg:
    to_display = [msg for msg in chat_history if msg['role'] == 'user' or msg['role'] == 'assistant']
else:
    logger.warning("Showing system messages")
    to_display = chat_history

# chat_txt =  "<br>".join(
#     [
#         f"<p class=assistant-msg>{msg['content']}</p>" if msg['role'] == 'assistant' else f"<i>{msg['content']}</i>"
#         for msg in to_display
#     ]
# )
scrollable_container = make_chat_container(to_display)
# print("Chat text:", scrollable_container)
chat_area.markdown(
    scrollable_container,
    unsafe_allow_html=True,
)

def clear_chat_history(chat_history):
    chat_history.clear()
    chat_history.append(dict(role="system", content=system_msg))

clear_btn = st.button("Clear Chat History", on_click=clear_chat_history, args=(chat_history,))
# No matter what the user message says, respond with "duck"

# logs
st.markdown("## Logs")
st.code(log_stream.getvalue())