chatlm-streamlit/requirements.txt
streamlit
numpy
pandas
openai
python-dotenv
markdown